--- resources/ui/gtk3/copyValueButton.ui
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-SPDX-FileCopyrightText: Simon Schneegans <code@simonschneegans.de>
-SPDX-License-Identifier: MIT
--->
-
-<interface domain="flypie">
-  <requires lib="gtk" version="3.0" />
-
-  <template class="FlyPieCopyValueButton" parent="GtkButton">
-    <property name="tooltip_text" translatable="yes">Copy left value to right side</property>
-    <property name="relief">none</property>
-    <child>
-      <object class="GtkImage">
-        <property name="icon_name">go-next-symbolic</property>
-      </object>
-    </child>
-    <style>
-      <class name="pill-button" />
-    </style>
-  </template>
-</interface>
