--- resources/ui/gtk3/iconSelectDialog.ui
@@ -1,144 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-SPDX-FileCopyrightText: Simon Schneegans <code@simonschneegans.de>
-SPDX-License-Identifier: MIT
--->
-
-<interface domain="flypie">
-  <requires lib="gtk" version="3.0" />
-
-  <template class="FlyPieIconSelectDialog" parent="GtkDialog">
-    <property name="modal">1</property>
-    <property name="title"></property>
-    <action-widgets>
-      <action-widget response="cancel">cancel</action-widget>
-      <action-widget response="ok" default="true">ok</action-widget>
-    </action-widgets>
-    <child type="action">
-      <object class="GtkButton" id="cancel">
-        <property name="label" translatable="yes">Cancel</property>
-      </object>
-    </child>
-    <child type="action">
-      <object class="GtkButton" id="ok">
-        <property name="label" translatable="yes">Select Icon</property>
-      </object>
-    </child>
-
-    <child internal-child="headerbar">
-      <object class="GtkBox">
-        <child type="title">
-          <object class="GtkStackSwitcher">
-            <property name="halign">center</property>
-            <property name="stack">stack</property>
-          </object>
-        </child>
-      </object>
-    </child>
-
-    <child internal-child="vbox">
-      <object class="GtkBox">
-        <child>
-
-          <object class="GtkStack" id="stack">
-            <property name="hexpand">1</property>
-            <property name="vexpand">1</property>
-            <property name="transition_duration">300</property>
-            <property name="transition_type">slide-left-right</property>
-
-            <child>
-              <object class="GtkBox">
-                <property name="orientation">vertical</property>
-
-                <child>
-                  <object class="GtkBox">
-                    <property name="halign">fill</property>
-                    <style>
-                      <class name="toolbar" />
-                    </style>
-                    <child type="center">
-                      <object class="GtkSearchEntry" id="filterEntry"></object>
-                    </child>
-                    <child>
-                      <object class="GtkSpinner" id="spinner">
-                        <property name="margin_start">5</property>
-                        <property name="margin_end">5</property>
-                        <property name="margin_top">5</property>
-                        <property name="margin_bottom">5</property>
-                        <property name="active">1</property>
-                      </object>
-                    </child>
-                  </object>
-                </child>
-
-                <child>
-                  <object class="GtkSeparator"></object>
-                </child>
-
-                <child>
-                  <object class="GtkScrolledWindow">
-                    <property name="hexpand">1</property>
-                    <property name="vexpand">1</property>
-                    <child>
-                      <object class="GtkIconView" id="iconView">
-                        <property name="item_orientation">horizontal</property>
-                        <property name="model">iconListFiltered</property>
-                        <property name="row_spacing">0</property>
-                        <property name="column_spacing">0</property>
-                        <property name="tooltip_column">0</property>
-                        <property name="item_padding">3</property>
-                        <child>
-                          <object class="GtkCellRendererPixbuf">
-                            <property name="stock_size">5</property>
-                          </object>
-                          <attributes>
-                            <attribute name="icon-name">0</attribute>
-                          </attributes>
-                        </child>
-                      </object>
-                    </child>
-                  </object>
-                </child>
-              </object>
-              <packing>
-                <property name="name">icon-theme-page</property>
-                <property name="title" translatable="yes" comments="Page title of the icon-select-dialog.">Icon Theme</property>
-              </packing>
-            </child>
-
-            <child>
-              <object class="GtkFileChooserWidget" id="iconFileChooser">
-                <property name="height_request">500</property>
-                <property name="filter">image-file-filter</property>
-              </object>
-              <packing>
-                <property name="name">custom-icon-page</property>
-                <property name="title" translatable="yes" comments="Page title of the icon-select-dialog.">Custom Icon</property>
-              </packing>
-            </child>
-
-          </object>
-
-        </child>
-      </object>
-    </child>
-
-  </template>
-
-  <object class="GtkListStore" id="iconList">
-    <columns>
-      <column type="gchararray" />
-    </columns>
-  </object>
-  <object class="GtkTreeModelFilter" id="iconListFiltered">
-    <property name="child_model">iconList</property>
-  </object>
-  <object class="GtkFileFilter" id="image-file-filter">
-    <mime-types>
-      <mime-type>image/*</mime-type>
-    </mime-types>
-  </object>
-
-
-</interface>
