--- resources/ui/gtk3/imageChooserButton.ui
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-SPDX-FileCopyrightText: Simon Schneegans <code@simonschneegans.de>
-SPDX-License-Identifier: MIT
--->
-
-<interface domain="flypie">
-  <requires lib="gtk" version="3.0" />
-
-  <template class="FlyPieImageChooserButton" parent="GtkBox">
-    <child>
-      <object class="GtkButton" id="button">
-        <child>
-          <object class="GtkBox">
-            <property name="spacing">8</property>
-            <child>
-              <object class="GtkLabel" id="label">
-                <property name="hexpand">1</property>
-                <property name="halign">start</property>
-                <property name="ellipsize">middle</property>
-                <property name="label" translatable="yes">(None)</property>
-              </object>
-            </child>
-            <child>
-              <object class="GtkImage">
-                <property name="icon_name">document-open-symbolic</property>
-              </object>
-            </child>
-          </object>
-        </child>
-      </object>
-    </child>
-    <child>
-      <object class="GtkButton" id="resetButton">
-        <child>
-          <object class="GtkImage">
-            <property name="icon_name">edit-clear-all-symbolic</property>
-          </object>
-        </child>
-      </object>
-    </child>
-    <style>
-      <class name="linked" />
-    </style>
-  </template>
-</interface>
