+++ src/prefs/ImageChooserButton.js
@@ -11,12 +11,10 @@
 
 'use strict';
 
-const {GObject, Gtk} = imports.gi;
+import GObject from 'gi://GObject';
+import Gtk from 'gi://Gtk';
 
 const _ = imports.gettext.domain('flypie').gettext;
-
-const Me    = imports.misc.extensionUtils.getCurrentExtension();
-const utils = Me.imports.src.common.utils;
 
 //////////////////////////////////////////////////////////////////////////////////////////
 // Since the FileChooserButton from GTK3 is gone, we have to provide a similar          //
@@ -24,13 +22,13 @@ const utils = Me.imports.src.common.utils;
 // type is hard-coded.                                                                  //
 //////////////////////////////////////////////////////////////////////////////////////////
 
-function registerWidget() {
+export function registerWidget() {
 
   if (GObject.type_from_name('FlyPieImageChooserButton') == null) {
     // clang-format off
       GObject.registerClass({
         GTypeName: 'FlyPieImageChooserButton',
-        Template: `resource:///ui/${utils.gtk4() ? "gtk4" : "gtk3"}/imageChooserButton.ui`,
+        Template: `resource:///ui/gtk4/imageChooserButton.ui`,
         InternalChildren: ['button', 'label', 'resetButton'],
         Signals: {
           'file-set': {}
@@ -57,7 +55,7 @@ function registerWidget() {
           filter: fileFilter
         });
 
-        utils.boxAppend(this._dialog.get_content_area(), this._fileChooser);
+        this._dialog.get_content_area().append(this._fileChooser);
 
         this._dialog.connect('response', (dialog, id) => {
           if (id == Gtk.ResponseType.OK) {
@@ -68,13 +66,9 @@ function registerWidget() {
         });
 
         this._button.connect('clicked', (button) => {
-          this._dialog.set_transient_for(utils.getRoot(button));
+          this._dialog.set_transient_for(button.get_root());
 
-          if (utils.gtk4()) {
-            this._dialog.show();
-          } else {
-            this._dialog.show_all();
-          }
+          this._dialog.show();
 
           if (this._file != null) {
             this._fileChooser.set_file(this._file);
